Porting related changes.  Mostly include file related.
Lifelines/Includes/ll-porting.h

    new defines: key_ot_record, node_to_string, free_nodes, ARRSIZE,
    key_to_record, qkey_to_record, length_indiseq, create_list
    fix typo in fam_to_first_chil.

Lifelines/Includes/options.h

    declare keyflag.

Lifelines/Includes/xlat.h

    Declare xlat.c routines.

Lifeslines/Makefile.am

    Add translat.c and xlat.c to list of files to build

Lifelines/add.c

    Add includes: llpy-externs.h, translat.h

Lifelines/advedit.c
Lifelines/ask.c

    Add include: translat.h

Lifelines/askgedc.c

    Add include: sys_inc.h, cast return value from ARRSIZE

Lifelines/askprogram.c
Lifelines/interact.c
Lifelines/show.c

    Cast return value from ARRSIZE

Lifelines/browse.c
Lifelines/main.c
Lifelines/show.c

    Add includes: translat.h, options.h

Lifelines/choose.c
Lifelines/delete.c
Lifelines/edit.c
Lifelines/equaliso.c
Lifelines/lbrowse.c
Lifelines/merge.c
Lifelines/miscutils.c
Lifelines/newrecs.c
Lifelines/selectdb.c
Lifelines/swap.c
Lifelines/tandem.c
Lifelines/translat.c
Lifelines/valid.c

    Add include: translat.c

Lifelines/screen.c

    Add include: list.h

Lifelines/xlat.c

    New file.

+++ Lifelines/askgedc.c
@@ -22,6 +22,7 @@
 #include "ll-porting.h"
 #include "standard.h"
 #include "llnls.h"
+#include "sys_inc.h"
 
 #include "rfmt.h"
 #include "sequence.h"
@@ -73,7 +74,7 @@ select_gedcoms (const struct dirent *entry)
 	CNSTRING goodexts[] = { ".ged", ".gedcom" };
 	INT i;
 	/* examine end of entry->d_name */
-	for (i=0; i<ARRSIZE(goodexts); ++i) {
+	for (i=0; i<(int)ARRSIZE(goodexts); ++i) {
 		CNSTRING entext = entry->d_name + strlen(entry->d_name) - strlen(goodexts[i]);
 		/* is it what we want ? use platform correct comparison, from path.c */
 		if (path_match(goodexts[i], entext))
@@ -115,14 +116,14 @@ add_gedcom_props (TABLE fileprops)
 		goto end_add_program_props;
 
 	/* initialize array where we record metainfo we want */
-	for (i=0; i<ARRSIZE(tagsfound); ++i)
+	for (i=0; i<(int)ARRSIZE(tagsfound); ++i)
 		tagsfound[i] = 0;
 
 	/* now read line by line looking for metainfo */
 	while (NULL != fgets(str, sizeof(str), fp) && str[strlen(str)-1]=='\n' && line<20) {
 		STRING p;
 		chomp(str); /* trim trailing CR or LF */
-		for (i=0; i<ARRSIZE(f_tags); ++i) {
+		for (i=0; i<(int)ARRSIZE(f_tags); ++i) {
 			CNSTRING tag = f_tags[i];
 			if (tagsfound[i])
 				continue; /* already have this tag */
@@ -145,7 +146,7 @@ add_gedcom_props (TABLE fileprops)
 	fclose(fp);
 
 	/* add any metainfo we found to the property table */
-	for (i=0; i<ARRSIZE(tagsfound); ++i) {
+	for (i=0; i<(int)ARRSIZE(tagsfound); ++i) {
 		if (tagsfound[i]) {
 			add_prop_dnum(fileprops, f_tags[i], tagsfound[i]);
 			strfree(&tagsfound[i]);
