+++ Lifelines/interact.c
@@ -190,7 +190,7 @@ translate_hdware_key (INT c)
 		, { KEY_ENTER, CMD_KY_ENTER }
 	};
 	int i;
-	for (i=0; i<ARRSIZE(hdkey); ++i) {
+	for (i=0; i<(int)ARRSIZE(hdkey); ++i) {
 		if (c == hdkey[i].key)
 			return hdkey[i].cmd;
 	}
@@ -208,7 +208,7 @@ translate_control_key (INT c)
 		, { '\n', CMD_KY_ENTER } /* UNIX */
 	};
 	int i;
-	for (i=0; i<ARRSIZE(hdkey); ++i) {
+	for (i=0; i<(int)ARRSIZE(hdkey); ++i) {
 		if (c == hdkey[i].key)
 			return hdkey[i].cmd;
 	}
